api_gestion_del_personal/pyproject.toml
[project]
name = "api-gestion-del-personal"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastmcp>=2.12.3",
    "httpx>=0.28.1",
    "mcp[cli]>=1.14.1",
    "python-dotenv",
]
